+++ AccountsWebApi/Controllers/EmployeesController.cs
@@ -8,6 +8,9 @@ using Microsoft.EntityFrameworkCore;
 using AccountsWebApi.Models;
 using static Microsoft.EntityFrameworkCore.DbLoggerCategory.Database;
 using System.Security.Cryptography;
+using JwtAuthenticationManager.Models;
+using JwtAuthenticationManager;
+using Microsoft.AspNetCore.Authorization;
 
 namespace AccountsWebApi.Controllers
 {
@@ -17,20 +20,28 @@ namespace AccountsWebApi.Controllers
     {
         private readonly UserDbContext _context;
         private readonly ILogger<EmployeesController> _logger;
-
-        public EmployeesController(UserDbContext context, ILogger<EmployeesController> logger)
+        private readonly JwtTokenHandler _JwtTokenHandler;
+        public EmployeesController(UserDbContext context, ILogger<EmployeesController> logger, JwtTokenHandler jwtTokenHandler)
         {
             _context = context;
             _logger = logger;
+            _JwtTokenHandler = jwtTokenHandler;
         }
 
         // GET: api/Employees
         [HttpGet]
-        public async Task<ActionResult<IEnumerable<Employee>>> Getemployees(string id)
+        [Authorize]
+        public async Task<ActionResult<IEnumerable<Employee>>> Getemployees()
         {
             try
             {
-                return await _context.employees.Where(x => x.companyId == id).ToListAsync();
+                var accessToken = Request.Headers.Authorization.FirstOrDefault()?.Replace("Bearer ", "");
+                var claimresponse = _JwtTokenHandler.GetCustomClaims(new ClaimRequest { token = accessToken, controllerActionName = RouteData.Values["controller"] + "Controller." + base.ControllerContext.ActionDescriptor.ActionName });
+                if (claimresponse.isAuth == true)
+                {
+                    return await _context.employees.Where(x => x.companyId == claimresponse.companyId && x.status == "Active").ToListAsync();
+                }
+                return Unauthorized();
             }
             catch(Exception ex)
             {
@@ -50,19 +61,26 @@ namespace AccountsWebApi.Controllers
         //}
 
         // GET: api/Employees/5
-        [HttpGet("{id}")]
-        public async Task<ActionResult<Employee>> GetEmployee(string id, string eId)
+        [HttpGet("getEmployee")]
+        [Authorize]
+        public async Task<ActionResult<Employee>> GetEmployee(string eId)
         {
             try
             {
-                var employee = await _context.employees.Where(x => x.companyId == id && x.employeeId == eId).FirstOrDefaultAsync();
-
-                if (employee == null)
+                var accessToken = Request.Headers.Authorization.FirstOrDefault()?.Replace("Bearer ", "");
+                var claimresponse = _JwtTokenHandler.GetCustomClaims(new ClaimRequest { token = accessToken, controllerActionName = RouteData.Values["controller"] + "Controller." + base.ControllerContext.ActionDescriptor.ActionName });
+                if (claimresponse.isAuth == true)
                 {
-                    return NotFound();
-                }
+                    var employee = await _context.employees.Where(x => x.companyId == claimresponse.companyId && x.employeeId == eId).FirstOrDefaultAsync();
 
-                return employee;
+                    if (employee == null)
+                    {
+                        return NotFound();
+                    }
+
+                    return employee;
+                }
+                return Unauthorized();
             }
             catch(Exception ex)
             {
@@ -74,24 +92,25 @@ namespace AccountsWebApi.Controllers
         // PUT: api/Employees/5
         // To protect from overposting attacks, see https://go.microsoft.com/fwlink/?linkid=2123754
         [HttpPut("{id}")]
-        public async Task<IActionResult> PutEmployee(string id, Employee employee)
+        [Authorize]
+        public async Task<IActionResult> PutEmployee(Employee employee)
         {
             try
             {
-                if (EmployeeExists(employee.employeeAutoId) == true)
+                var accessToken = Request.Headers.Authorization.FirstOrDefault()?.Replace("Bearer ", "");
+                var claimresponse = _JwtTokenHandler.GetCustomClaims(new ClaimRequest { token = accessToken, controllerActionName = RouteData.Values["controller"] + "Controller." + base.ControllerContext.ActionDescriptor.ActionName });
+                if (claimresponse.isAuth == true)
                 {
-                    if (id != employee.employeeId)
+                    if (EmployeeExists(employee.employeeAutoId))
                     {
-                        return BadRequest();
+                        employee.companyId = claimresponse.companyId;
+                        _context.Entry(employee).State = EntityState.Modified;
+
+                        await _context.SaveChangesAsync();
                     }
-
-                    _context.Entry(employee).State = EntityState.Modified;
-
-                    
-                    await _context.SaveChangesAsync();
-                    
+                    return NoContent();
                 }
-                return NoContent();
+                return Unauthorized();
             }
             catch(Exception ex)
             {
@@ -103,55 +122,61 @@ namespace AccountsWebApi.Controllers
         // POST: api/Employees
         // To protect from overposting attacks, see https://go.microsoft.com/fwlink/?linkid=2123754
         [HttpPost]
+        [Authorize]
         public async Task<ActionResult<IEnumerable<Employee>>> PostEmployee(Employee employee)
         {
             try
             {
-
-                var compId = _context.employees.Where(e => e.companyId == employee.companyId).Select(e => e.employeeId.ToString()).ToList();
-                var autoId = "";
-                if (compId.Count> 0)
+                var accessToken = Request.Headers.Authorization.FirstOrDefault()?.Replace("Bearer ", "");
+                var claimresponse = _JwtTokenHandler.GetCustomClaims(new ClaimRequest { token = accessToken, controllerActionName = RouteData.Values["controller"] + "Controller." + base.ControllerContext.ActionDescriptor.ActionName });
+                if (claimresponse.isAuth == true)
                 {
-                    autoId = compId.Max(x => int.Parse(x.Substring(1))).ToString();
-                }
 
-                if (autoId == "")
-                {
-                    _context.ChangeTracker.Clear();
-                    Employee c = new Employee();
-                    string comId = "E1";
-                    c.employeeId = comId;
-                    c.employeeName = employee.employeeName;
-                    c.employeeeMail = employee.employeeeMail;
-                    c.employeeContactNo = employee.employeeContactNo;
-                    c.employeeDesignation = employee.employeeDesignation;
-                    c.employeeFatherName = employee.employeeFatherName;
-                    c.deptAutoId = employee.deptAutoId;
-                    c.companyId = employee.companyId;
-                    c.status = employee.status;
-                    _context.employees.Add(c);
-                    await _context.SaveChangesAsync();
-                }
-                if (autoId != "")
-                {
-                    _context.ChangeTracker.Clear();
-                    Employee c = new Employee();
-                    string comId = "E" + (int.Parse(autoId) + 1);
-                    c.employeeId = comId;
-                    c.employeeName = employee.employeeName;
-                    c.employeeeMail = employee.employeeeMail;
-                    c.employeeContactNo = employee.employeeContactNo;
-                    c.employeeDesignation = employee.employeeDesignation;
-                    c.employeeFatherName = employee.employeeFatherName;
-                    c.deptAutoId = employee.deptAutoId;
-                    c.companyId = employee.companyId;
-                    c.status = employee.status;
-                    _context.employees.Add(c);
-                    await _context.SaveChangesAsync();
-                }
+                    var compId = _context.employees.Where(e => e.companyId == claimresponse.companyId).Select(e => e.employeeId.ToString()).ToList();
+                    var autoId = "";
+                    if (compId.Count > 0)
+                    {
+                        autoId = compId.Max(x => int.Parse(x.Substring(1))).ToString();
+                    }
 
-                return Ok();
+                    if (autoId == "")
+                    {
+                        _context.ChangeTracker.Clear();
+                        Employee c = new Employee();
+                        string comId = "E1";
+                        c.employeeId = comId;
+                        c.employeeName = employee.employeeName;
+                        c.employeeeMail = employee.employeeeMail;
+                        c.employeeContactNo = employee.employeeContactNo;
+                        c.employeeDesignation = employee.employeeDesignation;
+                        c.employeeFatherName = employee.employeeFatherName;
+                        c.deptAutoId = employee.deptAutoId;
+                        c.companyId = claimresponse.companyId;
+                        c.status = employee.status;
+                        _context.employees.Add(c);
+                        await _context.SaveChangesAsync();
+                    }
+                    if (autoId != "")
+                    {
+                        _context.ChangeTracker.Clear();
+                        Employee c = new Employee();
+                        string comId = "E" + (int.Parse(autoId) + 1);
+                        c.employeeId = comId;
+                        c.employeeName = employee.employeeName;
+                        c.employeeeMail = employee.employeeeMail;
+                        c.employeeContactNo = employee.employeeContactNo;
+                        c.employeeDesignation = employee.employeeDesignation;
+                        c.employeeFatherName = employee.employeeFatherName;
+                        c.deptAutoId = employee.deptAutoId;
+                        c.companyId = claimresponse.companyId;
+                        c.status = employee.status;
+                        _context.employees.Add(c);
+                        await _context.SaveChangesAsync();
+                    }
 
+                    return Ok();
+                }
+                return Unauthorized();
 
             }
             catch(Exception ex)
@@ -181,20 +206,31 @@ namespace AccountsWebApi.Controllers
 
         // DELETE: api/Employees/5
         [HttpDelete("{id}")]
-        public async Task<IActionResult> DeleteEmployee(string id, string eId)
+        [Authorize]
+        public async Task<IActionResult> DeleteEmployee(int id)
         {
             try
             {
-                var employee = _context.employees.Where(x => x.companyId == id && x.employeeId == eId).FirstOrDefault();
-                if (employee == null)
+                var accessToken = Request.Headers.Authorization.FirstOrDefault()?.Replace("Bearer ", "");
+                var claimresponse = _JwtTokenHandler.GetCustomClaims(new ClaimRequest { token = accessToken, controllerActionName = RouteData.Values["controller"] + "Controller." + base.ControllerContext.ActionDescriptor.ActionName });
+                if (claimresponse.isAuth == true)
                 {
+                    if (EmployeeExists(id))
+                    {
+                        var employee = await _context.employees.FindAsync(id);
+                        if (employee == null)
+                        {
+                            return NotFound();
+                        }
+
+                        _context.employees.Remove(employee);
+                        await _context.SaveChangesAsync();
+
+                        return Ok();
+                    }
                     return NotFound();
                 }
-
-                _context.employees.Remove(employee);
-                await _context.SaveChangesAsync();
-
-                return NoContent();
+                return Unauthorized();
             }
             catch(Exception ex)
             {
