+++ AccountsWebApi/Controllers/FacilitiesController.cs
@@ -8,6 +8,9 @@ using Microsoft.EntityFrameworkCore;
 using AccountsWebApi.Models;
 using static Microsoft.EntityFrameworkCore.DbLoggerCategory.Database;
 using System.Security.Cryptography;
+using JwtAuthenticationManager.Models;
+using JwtAuthenticationManager;
+using Microsoft.AspNetCore.Authorization;
 
 namespace AccountsWebApi.Controllers
 {
@@ -17,19 +20,28 @@ namespace AccountsWebApi.Controllers
     {
         private readonly UserDbContext _context;
         private readonly ILogger<FacilitiesController> _logger;
-        public FacilitiesController(UserDbContext context, ILogger<FacilitiesController> logger)
+        private readonly JwtTokenHandler _JwtTokenHandler;
+        public FacilitiesController(UserDbContext context, ILogger<FacilitiesController> logger, JwtTokenHandler jwtTokenHandler)
         {
             _context = context;
             _logger = logger;
+            _JwtTokenHandler = jwtTokenHandler;
         }
 
         // GET: api/Facilities
         [HttpGet]
-        public async Task<ActionResult<IEnumerable<Facility>>> Getfacilities(string cId)
+        [Authorize]
+        public async Task<ActionResult<IEnumerable<Facility>>> Getfacilities()
         {
             try
             {
-                return await _context.facilities.Where(x => x.companyId == cId).ToListAsync();
+                var accessToken = Request.Headers.Authorization.FirstOrDefault()?.Replace("Bearer ", "");
+                var claimresponse = _JwtTokenHandler.GetCustomClaims(new ClaimRequest { token = accessToken, controllerActionName = RouteData.Values["controller"] + "Controller." + base.ControllerContext.ActionDescriptor.ActionName });
+                if (claimresponse.isAuth == true)
+                {
+                    return await _context.facilities.Where(x => x.companyId == claimresponse.companyId && x.status == "Active" ).ToListAsync();
+                }
+                return Unauthorized();
             }
             catch(Exception ex)
             {
@@ -39,19 +51,27 @@ namespace AccountsWebApi.Controllers
         }
 
         // GET: api/Facilities/5
-        [HttpGet("{id}")]
-        public async Task<ActionResult<Facility>> GetFacility(int id)
+        [HttpGet("getFacility")]
+        [Authorize]
+        public async Task<ActionResult<Facility>> GetFacility(string id)
         {
             try
             {
-                var facility = await _context.facilities.FindAsync(id);
-
-                if (facility == null)
+                var accessToken = Request.Headers.Authorization.FirstOrDefault()?.Replace("Bearer ", "");
+                var claimresponse = _JwtTokenHandler.GetCustomClaims(new ClaimRequest { token = accessToken, controllerActionName = RouteData.Values["controller"] + "Controller." + base.ControllerContext.ActionDescriptor.ActionName });
+                if (claimresponse.isAuth == true)
                 {
-                    return NotFound();
-                }
+                    //var facility = await _context.facilities.FindAsync(id);
+                    var facility= await _context.facilities.Where(x => x.companyId == claimresponse.companyId).ToListAsync();
 
-                return facility;
+                    if (facility == null)
+                    {
+                        return NotFound();
+                    }
+
+                    return Ok(facility);
+                }
+                return Unauthorized();
             }
             catch(Exception ex)
             {
@@ -63,34 +83,24 @@ namespace AccountsWebApi.Controllers
         // PUT: api/Facilities/5
         // To protect from overposting attacks, see https://go.microsoft.com/fwlink/?linkid=2123754
         [HttpPut("{id}")]
+        [Authorize]
         public async Task<IActionResult> PutFacility(int id, Facility facility)
         {
             try
             {
-                if (id != facility.facilityAutoId)
+                var accessToken = Request.Headers.Authorization.FirstOrDefault()?.Replace("Bearer ", "");
+                var claimresponse = _JwtTokenHandler.GetCustomClaims(new ClaimRequest { token = accessToken, controllerActionName = RouteData.Values["controller"] + "Controller." + base.ControllerContext.ActionDescriptor.ActionName });
+                if (claimresponse.isAuth == true)
                 {
-                    return BadRequest();
-                }
 
-                _context.Entry(facility).State = EntityState.Modified;
+                    facility.companyId = claimresponse.companyId;
+                    _context.Entry(facility).State = EntityState.Modified;
 
-                try
-                {
+                    
                     await _context.SaveChangesAsync();
+                    return Ok();
                 }
-                catch (DbUpdateConcurrencyException)
-                {
-                    if (!FacilityExists(id))
-                    {
-                        return NotFound();
-                    }
-                    else
-                    {
-                        throw;
-                    }
-                }
-
-                return NoContent();
+                return Unauthorized();
             }
             catch(Exception ex)
             {
@@ -103,45 +113,52 @@ namespace AccountsWebApi.Controllers
         // POST: api/Facilities
         // To protect from overposting attacks, see https://go.microsoft.com/fwlink/?linkid=2123754
         [HttpPost]
+        [Authorize]
         public async Task<ActionResult<Facility>> PostFacility(Facility facility)
         {
             try
             {
-                var compId = _context.facilities.Where(d => d.companyId == facility.companyId).Select(d => d.facilityId).ToList();
-                var autoId = "";
-                if (compId.Count > 0)
+                var accessToken = Request.Headers.Authorization.FirstOrDefault()?.Replace("Bearer ", "");
+                var claimresponse = _JwtTokenHandler.GetCustomClaims(new ClaimRequest { token = accessToken, controllerActionName = RouteData.Values["controller"] + "Controller." + base.ControllerContext.ActionDescriptor.ActionName });
+                if (claimresponse.isAuth == true)
                 {
-                    autoId = compId.Max(x => int.Parse(x.Substring(1))).ToString();
-                }
+                    var compId = _context.facilities.Where(d => d.companyId == claimresponse.companyId).Select(d => d.facilityId).ToList();
+                    var autoId = "";
+                    if (compId.Count > 0)
+                    {
+                        autoId = compId.Max(x => int.Parse(x.Substring(1))).ToString();
+                    }
 
-                if (autoId == "")
-                {
-                    _context.ChangeTracker.Clear();
-                    Facility f = new Facility();
-                    string comId = "B1";
-                    f.facilityId = comId;
-                    f.facilityName = facility.facilityName;
-                    f.companyId = facility.companyId;
-                    f.status = facility.status;
-                    _context.facilities.Add(f);
-                    await _context.SaveChangesAsync();
-                    //return Ok(c);
-                }
-                if (autoId != "")
-                {
-                    _context.ChangeTracker.Clear();
-                    Facility f = new Facility();
-                    string comId = "B" + (int.Parse(autoId) + 1);
-                    f.facilityId = comId;
-                    f.facilityName = facility.facilityName;
-                    f.companyId = facility.companyId;
-                    f.status = facility.status;
-                    _context.facilities.Add(f);
-                    await _context.SaveChangesAsync();
-                    //return Ok(c);
-                }
+                    if (autoId == "")
+                    {
+                        _context.ChangeTracker.Clear();
+                        Facility f = new Facility();
+                        string comId = "B1";
+                        f.facilityId = comId;
+                        f.facilityName = facility.facilityName;
+                        f.companyId = claimresponse.companyId;
+                        f.status = facility.status;
+                        _context.facilities.Add(f);
+                        await _context.SaveChangesAsync();
+                        //return Ok(c);
+                    }
+                    if (autoId != "")
+                    {
+                        _context.ChangeTracker.Clear();
+                        Facility f = new Facility();
+                        string comId = "B" + (int.Parse(autoId) + 1);
+                        f.facilityId = comId;
+                        f.facilityName = facility.facilityName;
+                        f.companyId = claimresponse.companyId;
+                        f.status = facility.status;
+                        _context.facilities.Add(f);
+                        await _context.SaveChangesAsync();
+                        //return Ok(c);
+                    }
 
-                return Ok();
+                    return Ok();
+                }
+                return Unauthorized();
             }
             catch(Exception ex)
             {
@@ -152,20 +169,31 @@ namespace AccountsWebApi.Controllers
 
         // DELETE: api/Facilities/5
         [HttpDelete("{id}")]
+        [Authorize]
         public async Task<IActionResult> DeleteFacility(int id)
         {
             try
             {
-                var facility = await _context.facilities.FindAsync(id);
-                if (facility == null)
+                var accessToken = Request.Headers.Authorization.FirstOrDefault()?.Replace("Bearer ", "");
+                var claimresponse = _JwtTokenHandler.GetCustomClaims(new ClaimRequest { token = accessToken, controllerActionName = RouteData.Values["controller"] + "Controller." + base.ControllerContext.ActionDescriptor.ActionName });
+                if (claimresponse.isAuth == true)
                 {
+                    if (FacilityExists(id))
+                    {
+                        var facility = await _context.facilities.FindAsync(id);
+                        if (facility == null)
+                        {
+                            return NotFound();
+                        }
+
+                        _context.facilities.Remove(facility);
+                        await _context.SaveChangesAsync();
+
+                        return Ok();
+                    }
                     return NotFound();
                 }
-
-                _context.facilities.Remove(facility);
-                await _context.SaveChangesAsync();
-
-                return NoContent();
+                return Unauthorized();
             }
             catch(Exception ex)
             {
